refactor: Gemini 응답 처리 정리
